Fix server crashing when exporting storage config
